Move panic source into i386 directory
