+++ kernel/lib/assert.hpp
@@ -10,8 +10,6 @@
  */
 #pragma once
 
-#include <sys/panic.hpp>
-
 /**
  * @brief Creates an assertion which must be held to be true.
  * If the assertion passes, nothing happens and the kernel
