+++ tests/test-helpers/test-panic.hpp
@@ -1,0 +1,4 @@
+#pragma once
+#include <stdint.h>
+
+void panic(const char* msg, const char* file, uint32_t line, const char* func);
