+++ tests/test-lib/test-bitmap.cpp
@@ -9,6 +9,7 @@
  *
  */
 #include <catch2/catch.hpp>
+#include <test-helpers/test-panic.hpp>
 #include <lib/bitset.cpp>
 
 #define TEST_BITMAP_SIZE 4096 / (sizeof(size_t) * CHAR_BIT)
